REF: Rename pub/sub mocks module
